Adding dependencies for nova 1. python-nova, nova-compute-*, nova-common in superseded-packages. 2. Also created python-nova.control and python-nova.postinst(may not be required) 3. Added nova-common.dirs 4. Added nova-common.control
IMP: /etc/nova contents are not managed but
     nova.conf is getting managed by puppet
impacts:
 a)nova-common.postinst is changed
 b)nova-compute-qemu.postinst
 c)nova-compute.postinst

Status:
Packages are getting created and installed. Services are failing to start
Impact of above changes and missing dependencies which are not highlighted during install.

Solution:
Files can be managed by puppet if required and need to figure out all the run time depedencies for each sevice.